backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1
tensorflow==2.18.0
opencv-python==4.10.0.84
python-dotenv==1.0.0
transformers==4.35.2
scikit-learn==1.5.2
